chore: move types into single file
