+++ packages/visx-responsive/src/types.ts
@@ -1,0 +1,12 @@
+export type WithSizeProps = {
+  debounceTime?: number;
+  enableDebounceLeadingCall?: boolean;
+};
+
+export type WithSizeProvidedProps = {
+  width?: number;
+  screenWidth?: number;
+  height?: number;
+  initWidth?: number;
+  initHeight?: number;
+};
